+++ packages/adapter-interfaces/src/index.ts
@@ -1,5 +1,5 @@
-export type * from "./types";
 export * from "./types";
-export type * from "./adapters";
+export * from "./utils";
 
-export const tmp = "tmp";
+export type * from "./types";
+export type * from "./adapters";