+++ packages/kysely/src/keys/list.ts
@@ -1,0 +1,14 @@
+import { Kysely } from "kysely";
+import { Database } from "../db";
+
+export function list(db: Kysely<Database>) {
+  return async () => {
+    const keys = await db
+      .selectFrom("keys")
+      .where("revoked_at", "is", null)
+      .selectAll()
+      .execute();
+
+    return keys;
+  };
+}